+++ src/snowflake/snowpark/udtf.py
@@ -32,6 +32,7 @@ from snowflake.snowpark._internal.type_utils import (
 from snowflake.snowpark._internal.udf_utils import (
     TABLE_FUNCTION_PROCESS_METHOD,
     UDFColumn,
+    check_python_runtime_version,
     check_register_args,
     cleanup_failed_permanent_registration,
     create_python_udf_or_sp,
@@ -661,6 +662,7 @@ class UDTFRegistration:
             all_imports,
             all_packages,
             upload_file_stage_location,
+            custom_python_runtime_version_allowed,
         ) = resolve_imports_and_packages(
             self._session,
             TempObjectType.TABLE_FUNCTION,
@@ -676,6 +678,11 @@ class UDTFRegistration:
             statement_params=statement_params,
             skip_upload_on_content_match=skip_upload_on_content_match,
         )
+
+        if not custom_python_runtime_version_allowed:
+            check_python_runtime_version(
+                self._session._runtime_version_from_requirement
+            )
 
         raised = False
         try: