+++ tests/integ/test_udf.py
@@ -46,7 +46,7 @@ from snowflake.snowpark.exceptions import (
     SnowparkInvalidObjectNameException,
     SnowparkSQLException,
 )
-from snowflake.snowpark.functions import call_udf, col, count_distinct, pandas_udf, udf
+from snowflake.snowpark.functions import call_udf, col, pandas_udf, udf
 from snowflake.snowpark.types import (
     ArrayType,
     BinaryType,
@@ -56,6 +56,8 @@ from snowflake.snowpark.types import (
     FloatType,
     Geography,
     GeographyType,
+    Geometry,
+    GeometryType,
     IntegerType,
     MapType,
     PandasDataFrame,
@@ -84,6 +86,14 @@ def setup(session, resources_path):
     Utils.upload_to_stage(
         session, tmp_stage_name, test_files.test_udf_py_file, compress=False
     )
+
+
+@pytest.fixture(autouse=True)
+def clean_up(session):
+    session.clear_packages()
+    session.clear_imports()
+    session._runtime_version_from_requirement = None
+    yield
 
 
 def test_basic_udf(session):
@@ -819,6 +829,10 @@ def test_type_hints(session):
     def return_geography_dict_udf(g: Geography) -> Dict[str, str]:
         return g
 
+    @udf
+    def return_geometry_dict_udf(g: Geometry) -> Dict[str, str]:
+        return g
+
     df = session.create_dataframe([[1, 4], [2, 3]]).to_df("a", "b")
     Utils.check_answer(
         df.select(
@@ -841,6 +855,11 @@ def test_type_hints(session):
     Utils.check_answer(
         TestData.geography_type(session).select(return_geography_dict_udf("geo")),
         [Row('{\n  "coordinates": [\n    30,\n    10\n  ],\n  "type": "Point"\n}')],
+    )
+
+    Utils.check_answer(
+        TestData.geometry_type(session).select(return_geometry_dict_udf("geo")),
+        [Row('{\n  "coordinates": [\n    20,\n    81\n  ],\n  "type": "Point"\n}')],
     )
 
 
@@ -1200,6 +1219,18 @@ def test_udf_geography_type(session):
     )
 
 
+def test_udf_geometry_type(session):
+    def get_type(g):
+        return str(type(g))
+
+    geometry_udf = udf(get_type, return_type=StringType(), input_types=[GeometryType()])
+
+    Utils.check_answer(
+        TestData.geometry_type(session).select(geometry_udf(col("geo"))).collect(),
+        [Row("<class 'dict'>")],
+    )
+
+
 @pytest.mark.skipif(
     IS_IN_STORED_PROC, reason="Named temporary udf is not supported in stored proc"
 )
@@ -1380,158 +1411,6 @@ def test_udf_parallel(session):
     (not is_pandas_and_numpy_available) or IS_IN_STORED_PROC,
     reason="numpy and pandas are required",
 )
-def test_add_packages(session):
-    session.add_packages(["numpy==1.23.5", "pandas==1.5.3", "pandas==1.5.3"])
-    assert session.get_packages() == {
-        "numpy": "numpy==1.23.5",
-        "pandas": "pandas==1.5.3",
-    }
-
-    # dateutil is a dependency of pandas
-    def get_numpy_pandas_dateutil_version() -> str:
-        return f"{numpy.__version__}/{pandas.__version__}/{dateutil.__version__}"
-
-    udf_name = Utils.random_name_for_temp_object(TempObjectType.FUNCTION)
-    session.udf.register(get_numpy_pandas_dateutil_version, name=udf_name)
-    # don't need to check the version of dateutil, as it can be changed on the server side
-    assert (
-        session.sql(f"select {udf_name}()").collect()[0][0].startswith("1.23.5/1.5.3")
-    )
-
-    # only add numpy, which will overwrite the previously added packages
-    # so pandas will not be available on the server side
-    def is_pandas_available() -> bool:
-        try:
-            import pandas  # noqa: F401
-        except ModuleNotFoundError:
-            return False
-        return True
-
-    session.udf.register(
-        is_pandas_available, name=udf_name, replace=True, packages=["numpy"]
-    )
-    Utils.check_answer(session.sql(f"select {udf_name}()"), [Row(False)])
-
-    # with an empty list of udf-level packages
-    # it will still fail even if we have session-level packages
-    def is_numpy_available() -> bool:
-        try:
-            import numpy  # noqa: F401
-        except ModuleNotFoundError:
-            return False
-        return True
-
-    session.udf.register(is_numpy_available, name=udf_name, replace=True, packages=[])
-    Utils.check_answer(session.sql(f"select {udf_name}()"), [Row(False)])
-
-    session.clear_packages()
-
-    # add module objects
-    # but we can't register a udf with these versions
-    # because the server might not have them
-    resolved_packages = session._resolve_packages(
-        [numpy, pandas, dateutil], validate_package=False
-    )
-    assert f"numpy=={numpy.__version__}" in resolved_packages
-    assert f"pandas=={pandas.__version__}" in resolved_packages
-    assert f"python-dateutil=={dateutil.__version__}" in resolved_packages
-
-    session.clear_packages()
-
-
-def test_add_packages_with_underscore(session):
-    packages = ["spacy-model-en_core_web_sm", "typing_extensions"]
-    count = (
-        session.table("information_schema.packages")
-        .where(col("package_name").in_(packages))
-        .select(count_distinct("package_name"))
-        .collect()[0][0]
-    )
-    if count != len(packages):
-        pytest.skip("These packages with underscores are not available")
-
-    udf_name = Utils.random_name_for_temp_object(TempObjectType.FUNCTION)
-
-    @udf(name=udf_name, packages=packages)
-    def check_if_package_installed() -> bool:
-        try:
-            import spacy
-            import typing_extensions  # noqa: F401
-
-            spacy.load("en_core_web_sm")
-            return True
-        except Exception:
-            return False
-
-    Utils.check_answer(session.sql(f"select {udf_name}()").collect(), [Row(True)])
-
-
-@pytest.mark.skipif(
-    IS_IN_STORED_PROC, reason="Need certain version of datautil/pandas/numpy"
-)
-def test_add_packages_negative(session, caplog):
-    with pytest.raises(ValueError) as ex_info:
-        session.add_packages("python-dateutil****")
-    assert "InvalidRequirement" in str(ex_info)
-
-    with pytest.raises(ValueError) as ex_info:
-        session.add_packages("dateutil")
-    assert "Cannot add package dateutil" in str(ex_info)
-
-    with pytest.raises(ValueError) as ex_info:
-        with caplog.at_level(logging.WARNING):
-            # using numpy version 1.16.6 here because using any other version raises a
-            # ValueError for "non-existent python version in Snowflake" instead of
-            # "package is already added".
-            # In case this test fails in the future, choose a version of numpy which
-            # is supportezd by Snowflake using query:
-            #     select package_name, array_agg(version)
-            #     from information_schema.packages
-            #     where language='python' and package_name like 'numpy'
-            #     group by package_name;
-            session.add_packages("numpy", "numpy==1.16.6")
-    assert "is already added" in str(ex_info)
-
-    with pytest.raises(ValueError) as ex_info:
-        session.remove_package("python-dateutil")
-    assert "is not in the package list" in str(ex_info)
-
-    with pytest.raises(ValueError) as ex_info:
-        session.add_packages("xgboost==0.1.0")
-    assert "xgboost==0.1.0 because it is not available in Snowflake." in str(ex_info)
-
-    session.clear_packages()
-
-
-@pytest.mark.skipif(
-    (not is_pandas_and_numpy_available) or IS_IN_STORED_PROC,
-    reason="numpy and pandas are required",
-)
-def test_add_requirements(session, resources_path):
-    test_files = TestFiles(resources_path)
-    session.clear_packages()
-
-    session.add_requirements(test_files.test_requirements_file)
-    assert session.get_packages() == {
-        "numpy": "numpy==1.23.5",
-        "pandas": "pandas==1.5.3",
-    }
-
-    udf_name = Utils.random_name_for_temp_object(TempObjectType.FUNCTION)
-
-    @udf(name=udf_name)
-    def get_numpy_pandas_version() -> str:
-        return f"{numpy.__version__}/{pandas.__version__}"
-
-    Utils.check_answer(session.sql(f"select {udf_name}()"), [Row("1.23.5/1.5.3")])
-
-    session.clear_packages()
-
-
-@pytest.mark.skipif(
-    (not is_pandas_and_numpy_available) or IS_IN_STORED_PROC,
-    reason="numpy and pandas are required",
-)
 def test_udf_describe(session):
     def get_numpy_pandas_version(s: str) -> str:
         return f"{s}{numpy.__version__}/{pandas.__version__}"
@@ -1540,7 +1419,8 @@ def test_udf_describe(session):
         get_numpy_pandas_version, packages=["numpy", "pandas"]
     )
     describe_res = session.udf.describe(get_numpy_pandas_version_udf).collect()
-    assert [row[0] for row in describe_res] == [
+    actual_fields = [row[0] for row in describe_res]
+    expected_fields = [
         "signature",
         "returns",
         "language",
@@ -1552,7 +1432,14 @@ def test_udf_describe(session):
         "runtime_version",
         "packages",
         "installed_packages",
+        "is_aggregate",
     ]
+    # We use zip such that it is compatible regardless of UDAF is enabled or not on the merge gate accounts
+    for actual_field, expected_field in zip(actual_fields, expected_fields):
+        assert (
+            actual_field == expected_field
+        ), f"Actual: {actual_fields}, Expected: {expected_fields}"
+
     for row in describe_res:
         if row[0] == "packages":
             assert "numpy" in row[1] and "pandas" in row[1]
@@ -1706,6 +1593,7 @@ def test_pandas_udf_type_hints(session):
             ("datetime64[ns]",),
         ),
         (GeographyType, [["POINT(30 10)"]], (dict,), ("object",)),
+        (GeometryType, [["POINT(30 10)"]], (dict,), ("object",)),
         (MapType, [[{1: 2}]], (dict,), ("object",)),
     ],
 )
@@ -1837,6 +1725,7 @@ def test_pandas_udf_input_variant(session):
             ("datetime64[ns]",),
         ),
         (GeographyType, [["POINT(30 10)"]], (dict,), ("object",)),
+        (GeometryType, [["POINT(30 10)"]], (dict,), ("object",)),
         (MapType, [[{1: 2}]], (dict,), ("object",)),
     ],
 )
@@ -1854,7 +1743,7 @@ def test_pandas_udf_return_types(session, _type, data, expected_types, expected_
     )
     result_df = df.select(series_udf("a")).to_pandas()
     result_val = result_df.iloc[0][0]
-    if _type in (ArrayType, MapType, GeographyType):  # TODO: SNOW-573478
+    if _type in (ArrayType, MapType, GeographyType, GeometryType):  # TODO: SNOW-573478
         result_val = json.loads(result_val)
     assert isinstance(
         result_val, expected_types
@@ -2161,3 +2050,20 @@ def test_secure_udf(session):
     )
     ddl_sql = f"select get_ddl('function', '{echo.name}(int)')"
     assert "SECURE" in session.sql(ddl_sql).collect()[0][0]
+
+
+@pytest.mark.skipif(
+    (not is_pandas_and_numpy_available) or IS_IN_STORED_PROC,
+    reason="numpy and pandas are required",
+)
+@pytest.mark.parametrize(
+    "func", [numpy.min, numpy.sqrt, numpy.tan, numpy.sum, numpy.median]
+)
+def test_numpy_udf(session, func):
+    numpy_udf = udf(
+        func, return_type=DoubleType(), input_types=[DoubleType()], packages=["numpy"]
+    )
+    df = session.range(-5, 5).to_df("a")
+    Utils.check_answer(
+        df.select(numpy_udf("a")).collect(), [Row(func(i)) for i in range(-5, 5)]
+    )